+++ message_ix/reporting.py
@@ -1,5 +1,10 @@
 from ixmp.reporting import Reporter as IXMPReporter
 
 
-class MessageReporter(IXMPReporter):
-    pass
+class Reporter(IXMPReporter):
+    """MESSAGEix Reporter.
+
+    """
+    def __init__(self):
+        # TODO add MESSAGE_ix specific nodes from a file
+        super(Reporter, self).__init__(self)